Reduce memory usage and no longer block when loading images
Sets image bindings as asynchronous so that each one appear as soon as loaded and so that the UI does not block.
Updates image converter to allow to keep scaled down images in memory.
Explicitly calls the garbage collector when changing the image collection to reduce memory usage more reliably and faster.