Fix wallpaper error with some files when saving to jpeg and switch from bmp to jpeg
Some files could not be set as a wallpaper using the jpeg format, the Windows API call failed but there was no error.
It seems the file must include resolution properties, so the application now adds them if there are missing.
The wallpaper format is now jpeg instead of bmp because it was making pixelated wallpapers with small source images.