Fix file exception when saving after removing an image and adding it back
The file deletion was done last to have the destructive actions at the end, so that an error during the save may be recovered.
This caused an issue when an image was removed and added back in the same save: the file move failed because of the still existing file.